feat(switch): Allow --create with remote-only branches
Previously, `wt switch --create <branch>` would error if <branch> existed
on a remote but not locally.

This change updates the pre-creation check to only consider local
branches. This enables users to create a local worktree for a branch
that currently only exists on a remote. A new `local_branch_exists`
helper was added to differentiate local from remote branch checks.